refactor windowing into own file and change to FixedWindows
